+++ lib/Sublist3r/sublist3r.py
@@ -44,8 +44,21 @@ is_windows = sys.platform.startswith('win')
 
 # Console Colors
 if is_windows:
-    # use no terminal colors on windows
-    G = Y = B = R = W = G = Y = B = R = W = ''
+    # Windows deserve coloring too :D
+    G = '\033[92m'  # green
+    Y = '\033[93m'  # yellow
+    B = '\033[94m'  # blue
+    R = '\033[91m'  # red
+    W = '\033[0m'   # white
+    try:
+        import win_unicode_console , colorama
+        win_unicode_console.enable()
+        colorama.init()
+        #Now the unicode will work ^_^
+    except:
+        print("[!] Error: Coloring libraries not installed ,no coloring will be used [Check the readme]")
+        G = Y = B = R = W = G = Y = B = R = W = ''
+        
 
 else:
     G = '\033[92m'  # green
@@ -96,10 +109,10 @@ def write_file(filename, subdomains):
             f.write(subdomain + "\r\n")
 
 
-def subdomain_cmp(d1, d2):
-    """cmp function for subdomains d1 and d2.
+def subdomain_sorting_key(hostname):
+    """Sorting key for subdomains
 
-    This cmp function orders subdomains from the top-level domain at the right
+    This sorting key orders subdomains from the top-level domain at the right
     reading left, then moving '^' and 'www' to the top of their group. For
     example, the following list is sorted correctly:
 
@@ -116,24 +129,10 @@ def subdomain_cmp(d1, d2):
     ]
 
     """
-    d1 = d1.split('.')[::-1]
-    d2 = d2.split('.')[::-1]
-
-    val = 1 if d1 > d2 else (-1 if d1 < d2 else 0)
-    if ((len(d1) < len(d2)) and
-        (d1[-1] == 'www') and
-        (d1[: - 1] == d2[:len(d1) - 1])):
-        val = -1
-    elif ((len(d1) > len(d2)) and
-          (d2[-1] == 'www') and
-          (d1[:len(d2) - 1] == d2[: - 1])):
-        val = 1
-    elif d1[:-1] == d2[:-1]:
-        if d1[-1] == 'www':
-            val = -1
-        elif d2[-1] == 'www':
-            val = 1
-    return val
+    parts = hostname.split('.')[::-1]
+    if parts[-1] == 'www':
+        return parts[:-1], 1
+    return parts, 0
 
 
 class enumratorBase(object):
@@ -147,6 +146,12 @@ class enumratorBase(object):
         self.engine_name = engine_name
         self.silent = silent
         self.verbose = verbose
+        self.headers = {
+              'User-Agent': 'Mozilla/5.0 (Windows NT 6.1; WOW64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/58.0.3029.110 Safari/537.36',
+              'Accept': 'text/html,application/xhtml+xml,application/xml;q=0.9,image/webp,*/*;q=0.8',
+              'Accept-Language': 'en-US,en;q=0.8',
+              'Accept-Encoding': 'gzip',
+          } 
         self.print_banner()
 
     def print_(self, text):
@@ -160,17 +165,10 @@ class enumratorBase(object):
         return
 
     def send_req(self, query, page_no=1):
-        headers = {
-            'User-Agent': 'Mozilla/5.0 (X11; Ubuntu; Linux x86_64; rv:38.0) Gecko/20100101 Firefox/38.0',
-            'Accept': 'text/html,application/xhtml+xml,application/xml;q=0.9,*/*;q=0.8',
-            'Accept-Language': 'en-GB,en;q=0.5',
-            'Accept-Encoding': 'gzip, deflate',
-            'Connection': 'keep-alive',
-        }
 
         url = self.base_url.format(query=query, page_no=page_no)
         try:
-            resp = self.session.get(url, headers=headers, timeout=self.timeout)
+            resp = self.session.get(url, headers=self.headers, timeout=self.timeout)
         except Exception:
             resp = None
         return self.get_response(resp)
@@ -290,7 +288,7 @@ class GoogleEnum(enumratorBaseThreaded):
                 subdomain = urlparse.urlparse(link).netloc
                 if subdomain and subdomain not in self.subdomains and subdomain != self.domain:
                     if self.verbose:
-                        self.print_("%s%s: %s%s"%(R, self.engine_name, W, subdomain))
+                        self.print_("%s%s: %s%s" % (R, self.engine_name, W, subdomain))
                     self.subdomains.append(subdomain.strip())
         except Exception:
             pass
@@ -324,12 +322,12 @@ class YahooEnum(enumratorBaseThreaded):
         self.engine_name = "Yahoo"
         self.MAX_DOMAINS = 10
         self.MAX_PAGES = 0
-        super(YahooEnum, self).__init__(base_url, self.engine_name,domain, subdomains, q=q, silent=silent, verbose=verbose)
+        super(YahooEnum, self).__init__(base_url, self.engine_name, domain, subdomains, q=q, silent=silent, verbose=verbose)
         self.q = q
         return
 
     def extract_domains(self, resp):
-        link_regx2 = re.compile('<span class=" fz-15px fw-m fc-12th wr-bw.*?">(.*?)</span>')
+        link_regx2 = re.compile('<span class=" fz-.*? fw-m fc-12th wr-bw.*?">(.*?)</span>')
         link_regx = re.compile('<span class="txt"><span class=" cite fw-xl fz-15px">(.*?)</span>')
         links_list = []
         try:
@@ -389,7 +387,7 @@ class AskEnum(enumratorBaseThreaded):
                 subdomain = urlparse.urlparse(link).netloc
                 if subdomain not in self.subdomains and subdomain != self.domain:
                     if self.verbose:
-                        self.print_("%s%s: %s%s"%(R, self.engine_name, W, subdomain))
+                        self.print_("%s%s: %s%s" % (R, self.engine_name, W, subdomain))
                     self.subdomains.append(subdomain.strip())
         except Exception:
             pass
@@ -516,7 +514,7 @@ class BaiduEnum(enumratorBaseThreaded):
 class NetcraftEnum(enumratorBaseThreaded):
     def __init__(self, domain, subdomains=None, q=None, silent=False, verbose=True):
         subdomains = subdomains or []
-        self.base_url = 'http://searchdns.netcraft.com/?restriction=site+ends+with&host={domain}'
+        self.base_url = 'https://searchdns.netcraft.com/?restriction=site+ends+with&host={domain}'
         self.engine_name = "Netcraft"
         self.lock = threading.Lock()
         super(NetcraftEnum, self).__init__(self.base_url, self.engine_name, domain, subdomains, q=q, silent=silent, verbose=verbose)
@@ -525,15 +523,8 @@ class NetcraftEnum(enumratorBaseThreaded):
 
     def req(self, url, cookies=None):
         cookies = cookies or {}
-        headers = {
-            'User-Agent': 'Mozilla/5.0 (X11; Ubuntu; Linux x86_64; rv:38.0) Gecko/20100101 Firefox/40.0',
-            'Accept': 'text/html,application/xhtml+xml,application/xml;q=0.9,*/*;q=0.8',
-            'Accept-Language': 'en-GB,en;q=0.5',
-            'Accept-Encoding': 'gzip, deflate',
-        }
-
         try:
-            resp = self.session.get(url, headers=headers, timeout=self.timeout,cookies=cookies)
+            resp = self.session.get(url, headers=self.headers, timeout=self.timeout, cookies=cookies)
         except Exception as e:
             self.print_(e)
             resp = None
@@ -620,14 +611,8 @@ class DNSdumpster(enumratorBaseThreaded):
 
     def req(self, req_method, url, params=None):
         params = params or {}
-        headers = {
-            'User-Agent': 'Mozilla/5.0 (X11; Ubuntu; Linux x86_64; rv:38.0) Gecko/20100101 Firefox/40.0',
-            'Accept': 'text/html,application/xhtml+xml,application/xml;q=0.9,*/*;q=0.8',
-            'Accept-Language': 'en-GB,en;q=0.5',
-            'Accept-Encoding': 'gzip, deflate',
-            'Referer': 'https://dnsdumpster.com'
-        }
-
+        headers = dict(self.headers)
+        headers['Referer'] = 'https://dnsdumpster.com'
         try:
             if req_method == 'GET':
                 resp = self.session.get(url, headers=headers, timeout=self.timeout)
@@ -686,15 +671,8 @@ class Virustotal(enumratorBaseThreaded):
 
     # the main send_req need to be rewritten
     def send_req(self, url):
-        headers = {
-            'User-Agent': 'Mozilla/5.0 (X11; Ubuntu; Linux x86_64; rv:38.0) Gecko/20100101 Firefox/40.0',
-            'Accept': 'text/html,application/xhtml+xml,application/xml;q=0.9,*/*;q=0.8',
-            'Accept-Language': 'en-GB,en;q=0.5',
-            'Accept-Encoding': 'gzip, deflate',
-        }
-
         try:
-            resp = self.session.get(url, headers=headers, timeout=self.timeout)
+            resp = self.session.get(url, headers=self.headers, timeout=self.timeout)
         except Exception as e:
             self.print_(e)
             resp = None
@@ -709,7 +687,7 @@ class Virustotal(enumratorBaseThreaded):
         return self.subdomains
 
     def extract_domains(self, resp):
-        link_regx = re.compile('<div class="enum.*?">.*?<a target="_blank" href=".*?">(.*?)</a>',re.S)
+        link_regx = re.compile('<div class="enum.*?">.*?<a target="_blank" href=".*?">(.*?)</a>', re.S)
         try:
             links = link_regx.findall(resp)
             for link in links:
@@ -735,15 +713,8 @@ class ThreatCrowd(enumratorBaseThreaded):
         return
 
     def req(self, url):
-        headers = {
-            'User-Agent': 'Mozilla/5.0 (X11; Ubuntu; Linux x86_64; rv:38.0) Gecko/20100101 Firefox/40.0',
-            'Accept': 'text/html,application/xhtml+xml,application/xml;q=0.9,*/*;q=0.8',
-            'Accept-Language': 'en-GB,en;q=0.5',
-            'Accept-Encoding': 'gzip, deflate',
-        }
-
         try:
-            resp = self.session.get(url, headers=headers, timeout=self.timeout)
+            resp = self.session.get(url, headers=self.headers, timeout=self.timeout)
         except Exception:
             resp = None
 
@@ -770,7 +741,7 @@ class ThreatCrowd(enumratorBaseThreaded):
                     continue
                 if subdomain not in self.subdomains and subdomain != self.domain:
                     if self.verbose:
-                        self.print_("%s%s: %s%s"%(R, self.engine_name, W, subdomain))
+                        self.print_("%s%s: %s%s" % (R, self.engine_name, W, subdomain))
                     self.subdomains.append(subdomain.strip())
         except Exception as e:
             pass
@@ -787,15 +758,8 @@ class CrtSearch(enumratorBaseThreaded):
         return
 
     def req(self, url):
-        headers = {
-            'User-Agent': 'Mozilla/5.0 (X11; Ubuntu; Linux x86_64; rv:38.0) Gecko/20100101 Firefox/40.0',
-            'Accept': 'text/html,application/xhtml+xml,application/xml;q=0.9,*/*;q=0.8',
-            'Accept-Language': 'en-GB,en;q=0.5',
-            'Accept-Encoding': 'gzip, deflate',
-        }
-
         try:
-            resp = self.session.get(url, headers=headers, timeout=self.timeout)
+            resp = self.session.get(url, headers=self.headers, timeout=self.timeout)
         except Exception:
             resp = None
 
@@ -816,9 +780,13 @@ class CrtSearch(enumratorBaseThreaded):
                 subdomain = link.strip()
                 if not subdomain.endswith(self.domain) or '*' in subdomain:
                     continue
+
+                if '@' in subdomain:
+                    subdomain = subdomain[subdomain.find('@')+1:]
+
                 if subdomain not in self.subdomains and subdomain != self.domain:
                     if self.verbose:
-                        self.print_("%s%s: %s%s"%(R, self.engine_name, W, subdomain))
+                        self.print_("%s%s: %s%s" % (R, self.engine_name, W, subdomain))
                     self.subdomains.append(subdomain.strip())
         except Exception as e:
             pass
@@ -833,16 +801,27 @@ class PassiveDNS(enumratorBaseThreaded):
         self.q = q
         super(PassiveDNS, self).__init__(base_url, self.engine_name, domain, subdomains, q=q, silent=silent, verbose=verbose)
         return
+    
+    def get_agent(self,ua=None):
+        agents_url = 'http://www.webuseragents.com/recent'
+        try:
+            resp = self.session.get(agents_url, headers=self.headers, timeout=self.timeout)
+            agents_list = self.get_response(resp)
+            agents_regex = re.compile('<a href="/ua/.*?>(.*)</a>')
+            agents = agents_regex.findall(agents_list)
+            ua =  random.choice(agents)        
+        except Exception as e:
+            pass
+        
+        return ua
 
     def req(self, url):
-        headers = {
-            'User-Agent': 'Mozilla/5.0 (Windows NT 6.1) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/41.0.2228.0 Safari/537.36',
-            'Accept': 'text/html,application/xhtml+xml,application/xml;q=0.9,*/*;q=0.8',
-            'Accept-Language': 'en-GB,en;q=0.5',
-            'Accept-Encoding': 'gzip, deflate',
-        }
-
         try:
+            headers = dict(self.headers)
+            user_agent = self.get_agent()
+            if user_agent:
+                headers['User-Agent'] = user_agent
+
             resp = self.session.get(url, headers=headers, timeout=self.timeout)
         except Exception as e:
             self.print_(e)
@@ -853,20 +832,20 @@ class PassiveDNS(enumratorBaseThreaded):
     def enumerate(self):
         url = self.base_url.format(domain=self.domain)
         resp = self.req(url)
+        if not resp:
+            return self.subdomains
+
         self.extract_domains(resp)
         return self.subdomains
 
     def extract_domains(self, resp):
-        link_regx = re.compile('<td>(.*?)</td>')
+        link_regx = re.compile('[a-zA-Z0-9.-]*\.'+ self.domain,re.IGNORECASE)
         try:
             links = link_regx.findall(resp)
-            for link in links:
-                if self.domain not in link:
-                    continue
-                subdomain = link[:link.find('[')].strip()
+            for subdomain in links:
                 if subdomain not in self.subdomains and subdomain != self.domain and subdomain.endswith(self.domain):
                     if self.verbose:
-                        self.print_("%s%s: %s%s" %( R, self.engine_name, W, subdomain))
+                        self.print_("%s%s: %s%s" % (R, self.engine_name, W, subdomain))
                     self.subdomains.append(subdomain.strip())
         except Exception:
             pass
@@ -927,7 +906,7 @@ def main(domain, threads, savefile, ports, silent, verbose, enable_bruteforce, e
     parsed_domain = urlparse.urlparse(domain)
 
     if not silent:
-        print(B + "[-] Enumerating subdomains now for %s"%parsed_domain.netloc + W)
+        print(B + "[-] Enumerating subdomains now for %s" % parsed_domain.netloc + W)
 
     if verbose and not silent:
         print(Y + "[-] verbosity is enabled, will show the subdomains results in realtime" + W)
@@ -985,10 +964,8 @@ def main(domain, threads, savefile, ports, silent, verbose, enable_bruteforce, e
     subdomains = search_list.union(bruteforce_list)
 
     if subdomains:
-        subdomains = sorted(
-            subdomains,
-            key=functools.cmp_to_key(subdomain_cmp),
-        )
+        subdomains = sorted(subdomains, key=subdomain_sorting_key)
+
         if savefile:
             write_file(savefile, subdomains)
 
@@ -997,7 +974,7 @@ def main(domain, threads, savefile, ports, silent, verbose, enable_bruteforce, e
 
         if ports:
             if not silent:
-                print(G + "[-] Start port scan now for the following ports: %s%s" % (Y,ports) + W)
+                print(G + "[-] Start port scan now for the following ports: %s%s" % (Y, ports) + W)
             ports = ports.split(',')
             pscan = portscan(subdomains, ports)
             pscan.run()