+++ CHANGELOG.md
@@ -2,10 +2,21 @@ Changelog:
 
 v0.2.0-dev:
 
-Add Gather Company Employee
+Add Gather Public Company Employee
 Add SVN Finder
 Update URL Request
 Rework Code
 Fix small bug
 Update Harvest Public Document Regex
 Add version for updater
+
+v0.2.1-dev:
+
+Add Belati Configuration Wizard - [Core]
+Add Database Support( On Progress )
+Rework Code
+Update Database System
+Update Creating Project Info
+Update Sublist3r
+Update Gather Company Plugin
+Update README